+++ Tests/BlueTriangleTests/Mocks/Mock.swift
@@ -5,10 +5,10 @@
 //  Copyright © 2021 Blue Triangle. All rights reserved.
 //
 
-import Foundation
-import Combine
-import XCTest
 @testable import BlueTriangle
+import Combine
+import Foundation
+import XCTest
 
 // swiftlint:disable line_length
 
@@ -273,159 +273,5 @@ extension Mock {
         """
         {\"pgTm\":2000000,\"AB\":\"MY_AB_TEST_ID\",\"btV\":\"\(sdkVersion)\",\"CN8\":8.8800000000000008,\"CN10\":10.1,\"CV13\":\"CV13\",\"siteID\":\"MY_SITE_ID\",\"CN9\":9.9900000000000002,\"CN18\":18.18,\"thisURL\":\"MY_URL\",\"pageType\":\"MY_PAGE_TYPE\",\"CN11\":11.109999999999999,\"campaign\":null,\"eventType\":9,\"CV14\":\"CV14\",\"CV1\":\"CV1\",\"CN19\":19.190000000000001,\"CN12\":12.119999999999999,\"CV2\":\"CV2\",\"maxCPU\":100,\"CV15\":\"CV15\",\"os\":\"iOS\",\"CmpS\":\"MY_CAMPAIGN_SOURCE\",\"txnName\":\"MY_SEGMENT_NAME\",\"CV3\":\"CV3\",\"minMemory\":10000000,\"EUOS\":\"\(os)\",\"sID\":999999999999999999,\"CN13\":13.130000000000001,\"CV4\":\"CV4\",\"CN20\":20.199999999999999,\"avgMemory\":50000000,\"CV5\":\"CV5\",\"CN1\":1.1100000000000001,\"CmpM\":\"MY_CAMPAIGN_MEDIUM\",\"CN14\":14.140000000000001,\"nst\":0,\"navigationType\":9,\"device\":\"Mobile\",\"CV6\":\"CV6\",\"CN2\":2.2200000000000002,\"avgCPU\":50,\"CV7\":\"CV7\",\"CV10\":\"CV10\",\"CN3\":3.3300000000000001,\"CmpN\":\"MY_CAMPAIGN_NAME\",\"unloadEventStart\":0,\"CN15\":15.15,\"CV8\":\"CV8\",\"CN4\":4.4400000000000004,\"wcd\":0,\"gID\":888888888888888888,\"RV\":1,\"CV9\":\"CV9\",\"RefURL\":\"MY_REFERRING_URL\",\"CN5\":5.5499999999999998,\"CV11\":\"CV11\",\"domInteractive\":1000000,\"CN16\":16.16,\"browser\":\"Native App\",\"minCPU\":1,\"bvzn\":\"Native App-\(appVersion)-\(os) \(osVersion)\",\"CN6\":6.6600000000000001,\"bv\":0.51,\"CV12\":\"CV12\",\"maxMemory\":100000000,\"CN7\":7.7699999999999996,\"CN17\":17.170000000000002,\"pageName\":\"MY_PAGE_NAME\",\"DCTR\":\"MY_DATA_CENTER\"}
         """
-    }
-}
-
-class LoggerMock: Logging {
-    var onInfo: (String) -> Void
-    var onError: (String) -> Void
-
-    init(onInfo: @escaping (String) -> Void = { _ in }, onError: @escaping (String) -> Void = { _ in }) {
-        self.onInfo = onInfo
-        self.onError = onError
-    }
-
-    func logInfo(_ message: @autoclosure () -> String, file: StaticString, function: StaticString, line: UInt) {
-        onInfo(message())
-    }
-
-    func logError(_ message: @autoclosure () -> String, file: StaticString, function: StaticString, line: UInt) {
-        onError(message())
-    }
-
-    func reset() {
-        onInfo = { _ in }
-        onError = { _ in }
-    }
-}
-
-class PerformanceMonitorMock: PerformanceMonitoring {
-    var report: PerformanceReport
-    var onStart: () -> Void
-    var onEnd: () -> Void
-    var measurementCount: Int = 10
-
-    init(
-        report: PerformanceReport = Mock.performanceReport,
-        onStart: @escaping () -> Void = { },
-        onEnd: @escaping () -> Void = { }
-    ) {
-        self.report = report
-        self.onStart = onStart
-        self.onEnd = onEnd
-    }
-
-    func start() {
-        onStart()
-    }
-
-    func end() {
-        onEnd()
-    }
-
-    func makeReport() -> PerformanceReport {
-        report
-    }
-
-    func reset() {
-        report = Mock.performanceReport
-        onStart = { }
-        onEnd = { }
-    }
-}
-
-struct ResourceUsageMock: ResourceUsageMeasuring {
-    static func cpu() -> Double {
-        0.25
-    }
-
-    static func memory() -> UInt64 {
-        100
-    }
-}
-
-class CaptureTimerManagerMock: CaptureTimerManaging {
-    var onStart: () -> Void
-    var onCancel: () -> Void
-    var handler: (() -> Void)?
-
-    init(
-        onStart: @escaping () -> Void = {},
-        onCancel: @escaping () -> Void = {},
-        handler: @escaping () -> Void  = {}
-    ) {
-        self.onStart = onStart
-        self.onCancel = onCancel
-        self.handler = handler
-    }
-
-    func start() {
-        onStart()
-    }
-
-    func cancel() {
-        onCancel()
-    }
-
-    func fireTimer() {
-        handler?()
-    }
-}
-
-struct UploaderMock: Uploading {
-    var onSend: (Request) -> Void = { _ in }
-
-    init(onSend: @escaping (Request) -> Void = { _ in }) {
-        self.onSend = onSend
-    }
-
-    func send(request: Request) {
-        onSend(request)
-    }
-}
-
-final class URLProtocolMock: URLProtocol {
-    static var responseQueue: DispatchQueue = .global()
-    static var responseDelay: TimeInterval? = 0.3
-    static var responseProvider: (URL) throws -> (Data, HTTPURLResponse) = { url in
-        (Data(), Mock.makeHTTPResponse(url: url))
-    }
-
-    override class func canInit(with request: URLRequest) -> Bool { true }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest { request }
-
-    override func startLoading() {
-        if let delay = Self.responseDelay {
-            guard client != nil else { return }
-            Self.responseQueue.asyncAfter(deadline: .now() + delay) {
-                self.respond()
-            }
-        } else {
-            respond()
-        }
-    }
-
-    override func stopLoading() { }
-
-    private func respond() {
-        guard let client = client else { return }
-        do {
-            let url = try XCTUnwrap(request.url)
-            let response = try Self.responseProvider(url)
-            client.urlProtocol(self, didReceive: response.1, cacheStoragePolicy: .notAllowed)
-            client.urlProtocol(self, didLoad: response.0)
-        } catch {
-            client.urlProtocol(self, didFailWithError: error)
-        }
-        client.urlProtocolDidFinishLoading(self)
-    }
-}
-
-extension URLSessionConfiguration {
-    static var mock: URLSessionConfiguration {
-        let config = URLSessionConfiguration.ephemeral
-        config.protocolClasses = [URLProtocolMock.self]
-        return config
     }
 }
